+++ api/ElBaul.Api.Common/Models/ClearPhotoDateBatchRequest.cs
@@ -1,0 +1,4 @@
+using ElBaul.Domain;
+namespace ElBaul.Api.Models;
+
+public record ClearPhotoDateBatchRequest(List<PhotoId> PhotoIds);